+++ backend/ddang/src/test/java/com/ddang/ddang/notification/application/util/NotificationPropertyTest.java
@@ -1,10 +1,10 @@
 package com.ddang.ddang.notification.application.util;
 
-import org.junit.jupiter.api.Test;
-
-import java.util.List;
-
 import static org.assertj.core.api.Assertions.assertThat;
+
+import com.ddang.ddang.notification.application.NotificationProperty;
+import java.util.List;
+import org.junit.jupiter.api.Test;
 
 class NotificationPropertyTest {
